docsage/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0

# ML / NLP
torch==2.3.0
transformers==4.41.0
tokenizers==0.19.1
sentence-transformers==3.0.0
datasets==2.19.0
accelerate==0.30.0

# Retrieval
faiss-cpu==1.8.0
rank-bm25==0.2.2
networkx==3.3

# Document parsing
pypdf2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.3.0
tabula-py==2.9.0

# Storage & caching
redis==5.0.4
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# Utilities
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.5.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
structlog==24.2.0
tenacity==8.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
pytest-cov==5.0.0